Implement TSE review.
